Support C99 standards (enums, assert)
Fix various errors when -Werror=pedantic flag is enabled:

Cast enum values to int to prevent errors like:
   deps/tinycbor/src/cborconstants_p.h:47:31:\
   error: ISO C restricts enumerator values to range of ‘int’\
   [-Werror=pedantic]
   MajorTypeMask           = ~0U << MajorTypeShift,

"error: ISO C99 does not support" ‘Static_assert’

This effort is needed for RIOT port of iotivity-contrained

Bug-RIOT: https://github.com/RIOT-OS/RIOT/issues/6241

+++ src/cborconstants_p.h
@@ -44,7 +44,7 @@ enum {
     IndefiniteLength        = 31U,
 
     MajorTypeShift          = SmallValueBitLength,
-    MajorTypeMask           = ~0U << MajorTypeShift,
+    MajorTypeMask           = (int) (~0U << MajorTypeShift),
 
     BreakByte               = (unsigned)Break | (SimpleTypesType << MajorTypeShift)
 };
